Add not-working/regex-misleading-character-class.ql
Add it as 'not-working' because it apparently sometimes reports wrong indices
and the patterns it finds would sometimes only be harder to read when adding
additional backslashes.
Could possibly be improved in the future by adding a predicate for getting a
range or character listed in a character class, but the amount of work
required for implementing this might not justify the benefit it gives.